# be/xmash/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Xmash
  description: Xmash
  version: 1.0.0
servers:
  - url: 'http'
paths:
  /ranking:
    get:
      summary: '순위 정보 조회'
      description: '순위 정보 목록을 조회합니다.'
      operationId: getRanking
      responses:
        '200':
          description: '순위 정보 목록을 성공적으로 조회했습니다'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RankingResponse'
        '500':
          description: Server error or unexpected issue.
      tags:
        - Ranking
  /game/{gameType}:
    post:
      summary: '게임 결과 등록'
      description: '게임 결과를 등록합니다'
      operationId: updateGameResult
      parameters:
        - name: gameType
          in: path
          description: '게임 타입을 지정합니다.'
          required: true
          schema:
            $ref: '#/components/schemas/GameType'
      requestBody:
        description: '게임 결과 등록 요청 데이터'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameRequest'
      responses:
        '200':
          description: '게임 결과 업데이트 완료 했습니다.'
          content:
            application/json:
              schema:
                type: boolean
        '500':
          description: '서버 오류 또는 예기치 않은 문제 발생.'
      tags:
        - Game

components:
  schemas:
    RankingResponse:
      type: object
      properties:
        userId:
          type: string
          description: '사용자의 고유 식별자입니다.'
        userName:
          type: string
          description: '사용자의 이름입니다.'
        lp:
          type: integer
          description: '사용자의 LP(랭킹 포인트)입니다.'
        rank:
          type: integer
          description: '사용자의 랭크입니다.'
        tier:
          type: string
          description: '사용자의 티어입니다.'
      required:
        - userId
        - userName
        - lp
        - rank
        - tier
    GameRequest:
      type: object
      properties:
        myScore:
          type: integer
          description: '내 점수'
          example: 15
        opponentScore:
          type: integer
          description: '상대 점수'
          example: 10
        opponentPlayerId:
          type: string
          description: '상대 플레이어 ID'
          example: 'player123'
    GameType:
      type: string
      enum:
        - NORMAL
      description: '게임 타입입니다'